CREST-177 CMN-15 Fix JWT compatibility across JDK 6/7/8
JWTs that are signed and/or encrypted on JDK 6/7 and then verified
and/or decrypted on JDK 8 will fail due to a change in the JDK 8 hash
algorithm, and vice versa.

In addition the JWT implementation did not take into account whitespace
differences in a JWT and would cause decryption of a JWE to fail due to
the assumption that whitespace is always present.